requirements.txt
kafka-python==2.0.2
jsonschema==4.21.1
faker==22.6.0
typer[all]==0.9.0
flask==3.0.2
flask-socketio==5.3.6
python-socketio==5.11.1 